+ Se vuelve a la version 47 donde no habia segmentations faults. Los agregados de las versiones actuales se iran agregando asegurando q no causen problemas de memoria.

+++ trunk/tp1/Geometria/Coordenadas.cpp
@@ -47,7 +47,7 @@ Coordenadas* Coordenadas::menorY(Coordenadas* c){
     return coordenadas;
 }
 
-float Coordenadas::distancia (Coordenadas* hasta){
+double Coordenadas::distancia (Coordenadas* hasta){
     int deltaX = hasta->getX() - this->x;
     int deltaY = hasta->getY() - this->y;
     return sqrt(deltaX*deltaX + deltaY*deltaY);
